LAPR5-179 #ai #implementation gateway changes

+++ g4s-spa-module/src/app/modules/login/login.component.ts
@@ -1,7 +1,9 @@
 import { Component, OnInit } from '@angular/core';
 import { FormBuilder, Validators } from '@angular/forms';
 import { Router } from '@angular/router';
+import { NgxSpinnerService } from 'ngx-spinner';
 import { environment } from 'src/environments/environment';
+import { PlayerService } from '../player/services/player.service';
 
 @Component({
   selector: 'app-login',
@@ -10,6 +12,10 @@ import { environment } from 'src/environments/environment';
 })
 export class LoginComponent implements OnInit {
 
+  loginError: boolean = false;
+
+  wrongLogin: boolean = false;
+
   loginForm = this.fb.group({
     email: [''],
     password: ['']
@@ -17,16 +23,43 @@ export class LoginComponent implements OnInit {
 
   constructor(
     private fb: FormBuilder,
-    private router: Router) { 
+    private pService: PlayerService,
+    private router: Router,
+    private spinner: NgxSpinnerService) { 
     }
 
   ngOnInit(): void {
   }
 
-  login(){
-    localStorage.clear();
-    localStorage.setItem('currentPlayer', this.loginForm.value.email);
-    this.router.navigate(['/get-feed']);
+  login(): void {
+    this.spinner.show();
+    this.loginError = false;
+    this.wrongLogin = false;
+    let email = this.loginForm.value.email;
+    let password = this.loginForm.value.password;
+    this.pService.login(email, password).subscribe({ next: data => {
+      let code = data;
+      console.log(code);
+      if(code == 1) {
+        localStorage.clear();
+        localStorage.setItem('currentPlayer', this.loginForm.value.email);
+        this.router.navigate(['/get-feed']);
+      } else {
+        this.wrongLogin = true;
+        console.log("Erro no login");
+        this.resetForm();
+      }
+      this.spinner.hide();
+    },
+      error: _error => {
+        this.loginError = true;
+        this.spinner.hide();
+      }
+    });
+  }
+
+  resetForm(): void {
+    this.loginForm.reset();
   }
 
   get f() { return this.loginForm.controls; }
